refactor: Split up serializers into types
