--- packages/@pollyjs/utils/src/utils/serializers/browser/index.js
@@ -1,3 +1,0 @@
-import serializeRequestBody from './request-body';
-
-export default { serializeRequestBody };
