--- packages/@pollyjs/utils/src/utils/serializers/node/index.js
@@ -1,3 +1,0 @@
-import serializeRequestBody from './request-body';
-
-export default { serializeRequestBody };
